Add any-nix-shell support
Why: By default, a nix development shell will open in bash. This plugin
allows one to use any shell (in my case fish) so that when a development
environment is launched (with `nix develop`) it will continue to use

How: Updating the fish init script and adding the any-nix-shell package
in the flake.nix.

Tags: fish, nix