Configure home manager to immutably manage VSCode extension
See: https://www.reddit.com/r/NixOS/comments/15mohek/installing_vscode_extensions_with_homemanager_not/

Why: There's an issue where VSCode appears to not load all of the
extensions that I've declared. It's possible that the culprit is that
the extension directory is being changed and is causing some kind of
fallback behavior. Not sure, to be honest. But worth trying this anyway.

How: Updating the vscode-config.nix file to make the extensions
directory immutable.

Tags: vscode, bug, nix, home-manager